add new consensus types crate
