+++ consensus/core/src/lib.rs
@@ -59,11 +59,9 @@ mod transaction_certifier;
 
 /// Exported consensus API.
 pub use authority_node::ConsensusAuthority;
-pub use block::{
-    BlockAPI, BlockRef, CertifiedBlock, CertifiedBlocksOutput, Round, TransactionIndex,
-};
+pub use block::{BlockAPI, CertifiedBlock, CertifiedBlocksOutput};
 /// Exported API for testing.
-pub use block::{BlockTimestampMs, TestBlock, Transaction, VerifiedBlock};
+pub use block::{TestBlock, Transaction, VerifiedBlock};
 pub use commit::{CommitDigest, CommitIndex, CommitRef, CommittedSubDag};
 pub use commit_consumer::{CommitConsumer, CommitConsumerMonitor};
 pub use context::Clock;
