+++ consensus/core/src/network/network_tests.rs
@@ -5,6 +5,7 @@ use std::{sync::Arc, time::Duration};
 
 use bytes::Bytes;
 use consensus_config::NetworkKeyPair;
+use consensus_types::block::Round;
 use futures::StreamExt as _;
 use parking_lot::Mutex;
 use rstest::rstest;
@@ -17,7 +18,6 @@ use super::{
 use crate::{
     block::{TestBlock, VerifiedBlock},
     context::Context,
-    Round,
 };
 
 trait ManagerBuilder {
