+++ consensus/core/src/storage/store_tests.rs
@@ -2,12 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use consensus_config::AuthorityIndex;
+use consensus_types::block::{BlockDigest, BlockRef};
 use rstest::rstest;
 use tempfile::TempDir;
 
 use super::{mem_store::MemStore, rocksdb_store::RocksDBStore, Store, WriteBatch};
 use crate::{
-    block::{BlockDigest, BlockRef, TestBlock, VerifiedBlock},
+    block::{TestBlock, VerifiedBlock},
     commit::{CommitDigest, TrustedCommit},
 };
