+++ consensus/types/README.md
@@ -1,0 +1,2 @@
+This crate contains types shared between Sui and Consensus.
+It should be kept with minimal dependencies.
