+++ crates/sui-core/src/consensus_types/consensus_output_api.rs
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 use std::{cmp::Ordering, fmt::Display};
 
-use consensus_core::{BlockAPI, BlockRef, CommitDigest, TransactionIndex, VerifiedBlock};
+use consensus_core::{BlockAPI, CommitDigest, VerifiedBlock};
+use consensus_types::block::{BlockRef, TransactionIndex};
 use sui_protocol_config::ProtocolConfig;
 use sui_types::{
     digests::ConsensusCommitDigest,
