+++ crates/sui-core/src/mock_consensus.rs
@@ -7,7 +7,7 @@ use crate::checkpoints::CheckpointServiceNoop;
 use crate::consensus_adapter::{BlockStatusReceiver, ConsensusClient, SubmitToConsensus};
 use crate::consensus_handler::SequencedConsensusTransaction;
 use crate::execution_scheduler::ExecutionSchedulerAPI;
-use consensus_core::BlockRef;
+use consensus_types::block::BlockRef;
 use prometheus::Registry;
 use std::sync::{Arc, Weak};
 use std::time::Duration;
