+++ crates/sui-core/src/unit_tests/consensus_tests.rs
@@ -12,7 +12,8 @@ use crate::checkpoints::CheckpointServiceNoop;
 use crate::consensus_handler::SequencedConsensusTransaction;
 use crate::execution_scheduler::ExecutionSchedulerAPI;
 use crate::mock_consensus::with_block_status;
-use consensus_core::{BlockRef, BlockStatus};
+use consensus_core::BlockStatus;
+use consensus_types::block::BlockRef;
 use fastcrypto::traits::KeyPair;
 use move_core_types::{account_address::AccountAddress, ident_str};
 use parking_lot::Mutex;
