+++ crates/sui-core/src/unit_tests/wait_for_effects_tests.rs
@@ -4,7 +4,7 @@
 use std::sync::Arc;
 use std::time::Duration;
 
-use consensus_core::{BlockRef, TransactionIndex};
+use consensus_types::block::{BlockRef, TransactionIndex};
 use fastcrypto::traits::KeyPair;
 use sui_test_transaction_builder::TestTransactionBuilder;
 use sui_types::base_types::{ObjectRef, SuiAddress, TransactionDigest};
