+++ crates/sui-core/src/wait_for_effects_request.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+use consensus_types::block::{BlockRef, TransactionIndex};
 use sui_types::{
     committee::EpochId,
     digests::{TransactionDigest, TransactionEffectsDigest},
