+++ crates/sui-types/Cargo.toml
@@ -20,7 +20,7 @@ byteorder.workspace = true
 bytes.workspace = true
 chrono.workspace = true
 consensus-config.workspace = true
-consensus-core.workspace = true
+consensus-types.workspace = true
 num_enum.workspace = true
 im.workspace = true
 itertools.workspace = true
